Extract modules for each type
